requirements.txt
tensorflow
numpy
streamlit
pillow
opencv-python
matplotlib
scikit-learn
